Get the Swiss Federal Statistical Office data catalog in a given language

This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.

bfs_get_catalog(language = c("de", "fr", "it", "en"))

Arguments

language
character The language of a BFS catalog.

Value

A data frame

See also

Examples

get_bfs_catalog(language = "de")

#> Error in get_bfs_catalog(language = "de"): impossible de trouver la fonction "get_bfs_catalog"