Get the BFS data catalog

This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.

bfs_get_catalog_data(language = "de")

Arguments

language

character The language of a BFS catalog.

Value

A data frame

A tbl_df (a type of data frame; see tibble or dplyr packages).

title

A character column with the title of the BFS dataset

language

A character column with the language of the BFS dataset

published

A character column with a number for the text from each Wikisource page downloaded

url_bfs

A character column with the URL of the related BFS webpage

url_px

A character column with the URL of the PX file

See also

Examples

bfs_get_catalog_data(language = "de")
#> GET request successful. Parsing...
#> # A tibble: 350 × 5
#>    title                                       language published url_bfs url_px
#>    <chr>                                       <chr>    <chr>     <chr>   <chr> 
#>  1 Flugbewegungen (Starts und Landungen) im L… de       Flugbewe… https:… https…
#>  2 Flugpassagiere im Linien- und Charterverke… de       Flugpass… https:… https…
#>  3 Produktions-, Auftrags- und Umsatzstatisti… de       Produkti… https:… https…
#>  4 Produktions-, Auftrags- und Umsatzstatisti… de       Produkti… https:… https…
#>  5 Produktions-, Auftrags- und Umsatzstatisti… de       Produkti… https:… https…
#>  6 Neue Inverkehrsetzungen von Strassenfahrze… de       Neue Inv… https:… https…
#>  7 Hotellerie: Ankünfte und Logiernächte der … de       Hoteller… https:… https…
#>  8 Hotellerie: Angebot und Nachfrage der geöf… de       Hoteller… https:… https…
#>  9 Hotellerie: Angebot und Nachfrage der geöf… de       Hoteller… https:… https…
#> 10 Hotellerie: Angebot und Nachfrage der geöf… de       Hoteller… https:… https…
#> # ℹ 340 more rows