Get the BFS tables catalog

Get the list of the tables available in the official RSS Feed of the Swiss Federal Statistical Office tables catalog.

Usage

bfs_get_catalog_tables(
  language = "de",
  title = NULL,
  spatial_division = NULL,
  prodima = NULL,
  inquiry = NULL,
  institution = NULL,
  publishing_year_start = NULL,
  publishing_year_end = NULL,
  order_nr = NULL,
  skip_limit = TRUE
)

Arguments

language

character The language of a BFS catalog, i.e. "de", "fr", "it" or "en".

title

character String to search in title, subtitle and supertitle

spatial_division

BFS datasets by spatial division, choose between "Switzerland", "Cantons", "Districts", "Communes", "Other spatial divisions" or "International"

prodima

numeric Get only specific BFS themes using one or multiple prodima numbers

inquiry

character BFS datasets for an inquiry

institution

character BFS datasets for an institution

publishing_year_start

character BFS datasets for a publishing year start

publishing_year_end

character BFS datasets for a publishing year end

order_nr

character Filter by BFS Number (FSO number)

skip_limit

boolean skip limit, TRUE or FALSE

Value

A data frame. Returns NULL if no connection.

A tbl_df (a type of data frame; see tibble or dplyr packages). Returns NULL if no connection.

title

A character column with the title of the BFS dataset

language

A character column with the language of the BFS dataset

publication_date

The published date of the BFS dataset in the tables catalog

number_asset

The BFS asset number

url_bfs

A character column with the URL of the related BFS webpage

url_table

A character column with the URL of the PX file

catalog_date

The released date of the current BFS tables catalog

See also

Examples

# \donttest{
bfs_get_catalog_tables(language = "en", title = "students", prodima = c(900212))
#> # A tibble: 5 × 7
#>   title              language publication_date    number_asset url_bfs url_table
#>   <chr>              <chr>    <dttm>                     <dbl> <chr>   <chr>    
#> 1 Students at unive… en       2024-05-01 09:02:00     31826381 https:… https://…
#> 2 Students at unive… en       2024-05-01 09:02:00     31826380 https:… https://…
#> 3 Students at unive… en       2024-03-28 08:30:00     31185431 https:… https://…
#> 4 Students at unive… en       2024-03-28 08:30:00     31185438 https:… https://…
#> 5 Students at unive… en       2024-03-28 08:30:00     31185427 https:… https://…
#> # ℹ 1 more variable: catalog_date <dttm>
# }