Get the comments/footnotes of a BFS dataset in a given language

Get the comments/footnotes of a BFS dataset in a given language

bfs_get_data_comments(
  url_bfs = NULL,
  language = "de",
  number_bfs = NULL,
  query = "all",
  clean_names = FALSE
)

Arguments

url_bfs

The URL page of a dataset.

language

Language of the dataset to be translated if exists.

number_bfs

The BFS number of a dataset.

query

a list with named values, a json query file or json query string using pxweb::pxweb_query().

clean_names

Clean column names using janitor::clean_names()

Value

A tbl_df (a type of data frame; see tibble or dplyr packages).

See also